+++ R/proteus.R
@@ -1,7 +1,0 @@
-#' proteus: multiform seq2seq model for time-feature analysis
-#'
-#' Seq2seq time-feature analysis based on variational model, with a wide range of distributions available for the latent variable.
-#'
-#' @docType package
-#' @name proteus
-"_PACKAGE"