+++ NEWS.md
@@ -1,3 +1,11 @@
 # proteus 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.
+
+
+# proteus 1.1.0
+* Automated estimation of differentiation variable (deriv).
+* Automated detection of cuda compliant device.
+* Various improvement and code consolidation.
+* Added function for hyper-param random search.
+* Added another loss function (score).